+++ app/assets/javascripts/ovto_app/actions.js.rb
@@ -9,6 +9,18 @@ class MyApp < Ovto::App
     include ProjectActions
     include TaskActions
 
+    def show_upcoming_tasks(state:)
+      return {main_view: MainViewInfo.new(type: :upcoming_tasks)}
+    end
+
+    def show_projects(state:)
+      return {main_view: MainViewInfo.new(type: :projects)}
+    end
+
+    def show_project(state:, project_id:)
+      return {main_view: MainViewInfo.new(type: :project, project_id: project_id)}
+    end
+
     def show_task_details(state:, task:)
       return {focused_task: task}
     end