+++ app/assets/javascripts/ovto_app/view.js.rb
@@ -1,5 +1,5 @@
 require 'ovto'
-require 'ovto_app/components/current_task_list'
+require 'ovto_app/components/main_content'
 require 'ovto_app/components/sidebar'
 require 'ovto_app/components/task_details'
 
@@ -8,11 +8,20 @@ class MyApp < Ovto::App
     def render(state:)
       o '.View' do
         o '.Main' do
-          o CurrentTaskList
+          o MainContent, main_view: state.main_view
           o Sidebar
         end
         if state.focused_task
           o TaskDetails, task: state.focused_task
+        end
+      end
+    end
+
+    # Only shown in mobile devices
+    class ShowProjectsLink < Ovto::Component
+      def render
+        o '.ShowProjectsLink', onclick: ->{ actions.show_projects() } do
+          o 'div', "Projects"
         end
       end
     end